int size fixes

+++ alsuq/include/alsuq/generator/STLMersenne.hpp
@@ -36,7 +36,7 @@ private:
     std::pair<std::mt19937_64, long>& generator;
     std::uniform_real_distribution<real> distribution{0.0, 1.0};
 
-    const size_t dimension;
+    const long dimension;
 };
 } // namespace generator
 } // namespace alsuq
